remove .mp4 in name on upload
the direct upload allows only .mp4-files. i'm not shure if the name-length is prechecked, but you do this anywhere on filetype-check (if mp4).

if the check is after the substring, it would be possible to fail there - so in some (anyway unallowed) cases, this could cause a unproper fail - but at all, this change would save work and make it less ugly.

btw-offtopic: nice sort-changes in next-list! would be really cool, when next video stays in category..