Adding of more translation
More translation and correction.

Note: There is a strange bug ... Some word can not be added in this file, the "charts" part of the site can not manage and does not display / graphics values.
Broken word not usable at the moment (are not included in this file :

$t['Queue Position'] = "Position dans la file d'attente";
$t['Today'] = "Aujourd'hui";
$t['Total Views (Today)'] = "Total des vues (aujourd'hui)";

Note2: In the state, the file does not allow "Admin" to see the graphics, but only the values. Everything works well for normal users.