Created Slackbot Plugin
This plugin is meant to communicate to a slack workspace as a bot.  It requires a special bot_user_oauth_access_token that can be generated in your slack workspace, with the required scopes.

This plugin iterates through the users who have subscribed to the current user’s channel.  On each iteration it takes the email from the subscribed user, and sends it to slack to look for a user id.  Then it takes the user id and uses it to send a message to the user via the slack bot channel in slack.

This plugin uses the users:read, and users:read:email scope to retrieve retrieve the slack id for the user