Always include 'content-type' header
Without this each download is marked a 'text/html', and downloaded as such. On my Mac that's not a problem, but when I download a file on an Android tablet, the resulting file is marked as a HTML file instead of as a video - and opening it will display a bunch of gibberish unicode instead of the video. This change includes the mime-type in each response, since (as far as I'm aware) there are no downsides to doing that.