Collect and store video metadata from ffprobe
A new videos_metadata table is created for this purpose, where
we store all informations returned by ffprobe.

The table is populated when encoder uploads the video. A
CLI script install/metadata.php is provided to load the
metadata for all existing videos.

A first usage is to display video resolution on the video download
button.