this-video-is-not-available.jpg
I created this new image. The original image has resolution and language issues.
It is not possible to change the language in the still image.
Another problem is that the same image appears when the search does not return any information. So I removed the saying "video".
That's my suggestion.