Files with UPPER case extensions not listing in bulk upload
This fix will take the same allowed list to be considered in it's current original case (can be mixed), lower case format and also upper case format. Since it is also possible that some formats are mixed case (Ex: Mov instead of MOV) it is better to take all three cases (though may have duplicates in $fileStr, does not hurt.).